+++ KekBot/Argument Resolvers/ChoicesConverter.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Converters;
 using DSharpPlus.Entities;
-using static KekBot.PickCommand;
+using static KekBot.Commands.PickCommand;
 
 namespace KekBot.ArgumentResolvers {
     class ChoicesConverter : IArgumentConverter<ChoicesList> {
