+++ KekBot/Program.cs
@@ -1,22 +1,22 @@
 ﻿using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using System.Threading.Tasks;
-using DSharpPlus;
-using DSharpPlus.CommandsNext;
-using DSharpPlus.Entities;
-using Newtonsoft.Json;
-using KekBot.ArgumentResolvers;
-using DSharpPlus.Interactivity;
-using KekBot.Command.Commands;
-using Newtonsoft.Json.Serialization;
 using System.Text.Json.Serialization;
-using System.Collections.Generic;
-using DSharpPlus.CommandsNext.Exceptions;
+using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
 using RethinkDb.Driver;
 using RethinkDb.Driver.Net;
-using System.Text.RegularExpressions;
+using DSharpPlus;
+using DSharpPlus.CommandsNext;
+using DSharpPlus.CommandsNext.Exceptions;
+using DSharpPlus.Entities;
+using DSharpPlus.Interactivity;
+using KekBot.ArgumentResolvers;
+using KekBot.Commands;
 using KekBot.Utils;
 
 namespace KekBot {
